fix  #3

+++ tools/clean.sh
@@ -8,8 +8,21 @@ for THEME_NANE in $( find $THEMES_PATH -maxdepth 1 -mindepth 1 -type d \( ! -ina
 do
 	THEME_NANE=$( basename $THEME_NANE )
 	outputH1 "Theme: '$THEME_NANE'"
-	rm "$THEMES_PATH/$THEME_NANE/gss/$TIMESTAMP_FNAME"
-	rm "$THEMES_PATH/$THEME_NANE/templates/$TIMESTAMP_FNAME"
-	rm -r "$WEB_THEMES_PATH/$THEME_NANE/css"
-	rm -r "$WEB_THEMES_PATH/$THEME_NANE/js"
+	if [ -f $THEMES_PATH/$THEME_NANE/gss/$TIMESTAMP_FNAME ]
+	then
+		rm "$THEMES_PATH/$THEME_NANE/gss/$TIMESTAMP_FNAME"
+	fi
+	if [ -f $THEMES_PATH/$THEME_NANE/templates/$TIMESTAMP_FNAME ]
+	then
+		rm "$THEMES_PATH/$THEME_NANE/templates/$TIMESTAMP_FNAME"
+	fi
+	if [ -d $WEB_THEMES_PATH/$THEME_NANE/css ]
+	then
+		rm -r "$WEB_THEMES_PATH/$THEME_NANE/css"
+	fi
+
+	if [ -d $WEB_THEMES_PATH/$THEME_NANE/js ]
+	then
+		rm -r "$WEB_THEMES_PATH/$THEME_NANE/js"
+	fi
 done
